Initial commit; partially failing tests
